test-pytest-sentry/pyproject.toml
[project]
name = "test"
version = "0"
requires-python = ">=3.13"

dependencies = [
    "ipdb>=0.13.13",
    "pytest-rerunfailures>=15.0",
    "pytest-sentry",
    "sentry-sdk==3.0.0a1",
]

[tool.uv.sources]
pytest-sentry = { path = "../../pytest-sentry", editable = true }
# sentry-sdk = { path = "../../sentry-python", editable = true }
